requirements.txt
Django==5.2.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.8.0
django-environ==0.12.0
psycopg[binary]==3.2.10
dj-database-url==3.0.1
whitenoise==6.11.0
djangorestframework-gis==1.2.0
weasyprint==66.0
